fix: 修复视频时长为浮点数时转换格式的问题

+++ learning-service/learning-course/src/main/java/cn/linter/learning/course/service/impl/ChapterServiceImpl.java
@@ -73,16 +73,15 @@ public class ChapterServiceImpl implements ChapterService {
     }
 
     private String transformVideoTime(String videoTime) {
-        long duration;
+        double duration;
         try {
-            duration = Long.parseLong(videoTime);
+            duration = Double.parseDouble(videoTime);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("参数异常");
         }
-        long minutes = duration / 60;
-        long seconds = duration % 60;
+        long minutes = (long) (duration / 60);
+        long seconds = (long) (duration % 60);
         return (minutes < 10 ? "0" : "") + minutes + ":" + (seconds < 10 ? "0" : "") + seconds;
     }
 
 }
-
